add credential foreign key

+++ api/app/postUpgrade.php
@@ -355,10 +355,14 @@ if(version_compare($version, '2.17.0','>=')) {
   }
   if(!Capsule::schema()->hasTable('user_credentials')) {
     require_once('database/UserCredential.php');
+  }
+  $user_credentials_foreign_keys = listTableForeignKeys('user_credentials');
+  if(Capsule::schema()->hasTable('user_credentials') && !in_array('user_credentials_user_id_foreign', $user_credentials_foreign_keys)) {
     Capsule::schema()->table('user_credentials', function ($table, $as = null, $connection = null) {
       $table->foreign('user_id')->references('user_id')->on('users')->onDelete('cascade')->onUpdate('cascade');
     });
   }
+  
   $setting = FrcPortal\Setting::firstOrCreate(['section' => 'notification', 'setting' => 'email_enable_smtp'], ['value' => false]);
   $setting = FrcPortal\Setting::firstOrCreate(['section' => 'notification', 'setting' => 'email_smtp_server'], ['value' => '']);
   $setting = FrcPortal\Setting::firstOrCreate(['section' => 'notification', 'setting' => 'email_smtp_port'], ['value' => '']);
